Struct fluent_uri::encoding::EStr

pub struct EStr<E: Encoder> { /* private fields */ }

Percent-encoded string slices.

Examples

Parse key-value pairs from a query string into a hash map:

use fluent_uri::encoding::{encoder::Query, EStr};
use std::collections::HashMap;

let query = "name=%E5%BC%A0%E4%B8%89&speech=%C2%A1Ol%C3%A9%21";
let map: HashMap<_, _> = EStr::<Query>::new(query)
    .split('&')
    .map(|s| s.split_once('=').unwrap_or((s, EStr::new(""))))
    .map(|(k, v)| (k.decode().into_string_lossy(), v.decode().into_string_lossy()))
    .collect();
assert_eq!(map["name"], "张三");
assert_eq!(map["speech"], "¡Olé!");

Implementations

impl<E: Encoder> EStr<E>

pub const fn new(s: &str) -> &Self

Converts a string slice to an EStr slice.

Only use this function when you have a percent-encoded string at hand. You may otherwise encode and append data to an EString which derefs to EStr.

Panics

Panics if the string is not properly encoded with E.

pub const fn try_new(s: &str) -> Option<&Self>

Converts a string slice to an EStr slice, returning None if the conversion fails.

pub fn as_str(&self) -> &str

Yields the underlying string slice.

pub fn len(&self) -> usize

Returns the length of the EStr slice in bytes.

pub fn is_empty(&self) -> bool

Checks whether the EStr slice is empty.

pub fn decode(&self) -> Decode<'_>

Decodes the EStr slice.

Note that this method will not decode U+002B (+) as 0x20 (space).

This method allocates only when there is any percent-encoded octet in the slice.

Panics

Panics at compile time if E::TABLE does not allow percent-encoding.

Examples

use fluent_uri::encoding::{encoder::Path, EStr};

let dec = EStr::<Path>::new("%C2%A1Hola%21").decode();
assert_eq!(dec.as_bytes(), &[0xc2, 0xa1, 0x48, 0x6f, 0x6c, 0x61, 0x21]);
assert_eq!(dec.into_string()?, "¡Hola!");

pub fn split(&self, delim: char) -> Split<'_, E>

Returns an iterator over subslices of the EStr slice separated by the given delimiter.

Panics

Panics if the delimiter is not a reserved character.

Examples

use fluent_uri::encoding::{encoder::Path, EStr};

assert!(EStr::<Path>::new("a,b,c").split(',').eq(["a", "b", "c"]));
assert!(EStr::<Path>::new(",").split(',').eq(["", ""]));
assert!(EStr::<Path>::new("").split(',').eq([""]));

pub fn split_once(&self, delim: char) -> Option<(&Self, &Self)>

Splits the EStr slice on the first occurrence of the given delimiter and returns prefix before delimiter and suffix after delimiter.

Returns None if the delimiter is not found.

Panics

Panics if the delimiter is not a reserved character.

Examples

use fluent_uri::encoding::{encoder::Path, EStr};

assert_eq!(
    EStr::<Path>::new("foo;bar;baz").split_once(';'),
    Some((EStr::new("foo"), EStr::new("bar;baz")))
);

assert_eq!(EStr::<Path>::new("foo").split_once(';'), None);

pub fn rsplit_once(&self, delim: char) -> Option<(&Self, &Self)>

Splits the EStr slice on the last occurrence of the given delimiter and returns prefix before delimiter and suffix after delimiter.

Returns None if the delimiter is not found.

Panics

Panics if the delimiter is not a reserved character.

Examples

use fluent_uri::encoding::{encoder::Path, EStr};

assert_eq!(
    EStr::<Path>::new("foo;bar;baz").rsplit_once(';'),
    Some((EStr::new("foo;bar"), EStr::new("baz")))
);

assert_eq!(EStr::<Path>::new("foo").rsplit_once(';'), None);

impl EStr<Path>

Extension methods for the path component of URI reference.

pub fn is_absolute(&self) -> bool

Checks whether the path is absolute, i.e., starting with '/'.

pub fn is_rootless(&self) -> bool

Checks whether the path is rootless, i.e., not starting with '/'.