Module fluent_uri::encoding::table

Byte pattern tables from RFC 3986.

The predefined table constants in this module are documented with the ABNF notation of RFC 2234.

Structs

Table

A table determining the byte patterns allowed in a string.

Constants

ALPHA

ALPHA = A-Z / a-z

DIGIT

DIGIT = 0-9

FRAGMENT

fragment = *( pchar / “/” / “?” )

GEN_DELIMS

gen-delims = “:” / “/” / “?” / “#” / “[” / “]” / “@”

HEXDIG

HEXDIG = DIGIT / “A” / “B” / “C” / “D” / “E” / “F” / “a” / “b” / “c” / “d” / “e” / “f”

IPV_FUTURE

IPvFuture = “v” 1*HEXDIG “.” 1*( unreserved / sub-delims / “:” )

PATH

path = *( pchar / “/” )

PCHAR

pchar = unreserved / pct-encoded / sub-delims / “:” / “@”

QUERY

query = *( pchar / “/” / “?” )

REG_NAME

reg-name = *( unreserved / pct-encoded / sub-delims )

RESERVED

reserved = gen-delims / sub-delims

SCHEME

scheme = ALPHA *( ALPHA / DIGIT / “+” / “-” / “.” )

SEGMENT_NC

segment-nz-nc = 1*( unreserved / pct-encoded / sub-delims / “@” )

SUB_DELIMS

sub-delims = “!” / “$” / “&” / “’” / “(” / “)” / “*” / “+” / “,” / “;” / “=”

UNRESERVED

unreserved = ALPHA / DIGIT / “-” / “.” / “_” / “~”

USERINFO

userinfo = *( unreserved / pct-encoded / sub-delims / “:” )