fluent_uri::pct_enc

Module encoder

Percent-encoders for URI/IRI components.

Structs

Data

An encoder for URI data which preserves only unreserved characters and encodes the others.

Fragment

An encoder for URI fragment.

IData

An encoder for IRI data which preserves only unreserved characters and encodes the others.

IFragment

An encoder for IRI fragment.

IPath

An encoder for IRI path.

IQuery

An encoder for IRI query.

IRegName

An encoder for IRI registered name.

IUserinfo

An encoder for IRI userinfo.

Path

An encoder for URI path.

Port

An encoder for URI/IRI port.

Query

An encoder for URI query.

RegName

An encoder for URI registered name.

Userinfo

An encoder for URI userinfo.