fluent_uri::pct_enc

Struct EStr

pub struct EStr<E: Encoder> { /* private fields */ }

Percent-encoded string slices.

The owned counterpart of EStr is EString. See its documentation if you want to build a percent-encoded string from scratch.

Type parameter

The EStr<E> type is parameterized over a type E that implements Encoder. The associated constant E::TABLE of type Table specifies the byte patterns allowed in a string. In short, the underlying byte sequence of an EStr<E> slice can be formed by joining any number of the following byte sequences:

ch.encode_utf8(&mut [0; 4]) where E::TABLE.allows(ch).

[b'%', hi, lo] where E::TABLE.allows_pct_encoded() && hi.is_ascii_hexdigit() && lo.is_ascii_hexdigit().

Comparison

EStr slices are compared lexicographically by their byte values. Normalization is not performed prior to comparison.

Examples

Parse key-value pairs from a query string into a hash map:

use fluent_uri::{pct_enc::EStr, UriRef};
use std::collections::HashMap;

let s = "?name=%E5%BC%A0%E4%B8%89&speech=%C2%A1Ol%C3%A9%21";
let query = UriRef::parse(s)?.query().unwrap();
let map: HashMap<_, _> = query
    .split('&')
    .map(|s| s.split_once('=').unwrap_or((s, EStr::EMPTY)))
    .map(|(k, v)| (k.decode().to_string_lossy(), v.decode().to_string_lossy()))
    .collect();
assert_eq!(map["name"], "张三");
assert_eq!(map["speech"], "¡Olé!");

Implementations

impl<E: Encoder> EStr<E>

pub const EMPTY: &'static Self

An empty EStr slice.

pub const fn new_or_panic(s: &str) -> &Self

Converts a string slice to an EStr slice.

Panics

Panics if the string is not properly encoded with E. For a non-panicking variant, use new.

pub const fn new(s: &str) -> Option<&Self>

Converts a string slice to an EStr slice, returning None if the conversion fails.

pub fn encode_byte(x: u8) -> &'static Self

Creates an EStr slice containing a single percent-encoded octet representing the given byte.

Panics

Panics at compile time if E::TABLE does not allow percent-encoded octets.

Examples

use fluent_uri::pct_enc::{encoder::Path, EStr};

assert_eq!(EStr::<Path>::encode_byte(b'1'), "%31");

pub fn as_str(&self) -> &str

Yields the underlying string slice.

pub fn len(&self) -> usize

Returns the length of the EStr slice in bytes.

pub fn is_empty(&self) -> bool

Checks whether the EStr slice is empty.

pub fn decode(&self) -> Decode<'_>

Returns an iterator used to decode the EStr slice.

Always split before decoding, as otherwise the data may be mistaken for component delimiters.

Note that the iterator will not decode U+002B (+) as 0x20 (space).

Panics

Panics at compile time if E::TABLE does not allow percent-encoded octets.

Examples

use fluent_uri::pct_enc::{encoder::Path, EStr};

let dec = EStr::<Path>::new_or_panic("%C2%A1Hola%21").decode();
assert_eq!(*dec.clone().to_bytes(), [0xc2, 0xa1, 0x48, 0x6f, 0x6c, 0x61, 0x21]);
assert_eq!(dec.to_string().unwrap(), "¡Hola!");

pub fn split(&self, delim: char) -> Split<'_, E>

Returns an iterator over subslices of the EStr slice separated by the given delimiter.

Panics

Panics if the delimiter is not a reserved character.

Examples

use fluent_uri::pct_enc::{encoder::Path, EStr};

assert!(EStr::<Path>::new_or_panic("a,b,c").split(',').eq(["a", "b", "c"]));
assert!(EStr::<Path>::new_or_panic(",").split(',').eq(["", ""]));
assert!(EStr::<Path>::EMPTY.split(',').eq([""]));

pub fn split_once(&self, delim: char) -> Option<(&Self, &Self)>

Splits the EStr slice on the first occurrence of the given delimiter and returns prefix before delimiter and suffix after delimiter.

Returns None if the delimiter is not found.

Panics

Panics if the delimiter is not a reserved character.

Examples

use fluent_uri::pct_enc::{encoder::Path, EStr};

assert_eq!(
    EStr::<Path>::new_or_panic("foo;bar;baz").split_once(';'),
    Some((EStr::new_or_panic("foo"), EStr::new_or_panic("bar;baz")))
);

assert_eq!(EStr::<Path>::new_or_panic("foo").split_once(';'), None);

pub fn rsplit_once(&self, delim: char) -> Option<(&Self, &Self)>

Splits the EStr slice on the last occurrence of the given delimiter and returns prefix before delimiter and suffix after delimiter.

Returns None if the delimiter is not found.

Panics

Panics if the delimiter is not a reserved character.

Examples

use fluent_uri::pct_enc::{encoder::Path, EStr};

assert_eq!(
    EStr::<Path>::new_or_panic("foo;bar;baz").rsplit_once(';'),
    Some((EStr::new_or_panic("foo;bar"), EStr::new_or_panic("baz")))
);

assert_eq!(EStr::<Path>::new_or_panic("foo").rsplit_once(';'), None);

impl<E: PathEncoder> EStr<E>

Extension methods for the path component.

pub fn is_absolute(&self) -> bool

Checks whether the path is absolute, i.e., starting with '/'.

pub fn is_rootless(&self) -> bool

Checks whether the path is rootless, i.e., not starting with '/'.

pub fn segments_if_absolute(&self) -> Option<Split<'_, E>>

Returns an iterator over the path segments, separated by '/'.

Returns None if the path is rootless. Use split instead if you need to split a rootless path on occurrences of '/'.

Note that the path can be empty when authority is present, in which case this method will return None.

Examples

use fluent_uri::Uri;

// Segments are separated by '/'.
// The empty string before a leading '/' is not a segment.
// However, segments can be empty in the other cases.
let path = Uri::parse("file:///path/to//dir/")?.path();
assert_eq!(path, "/path/to//dir/");
assert!(path.segments_if_absolute().unwrap().eq(["path", "to", "", "dir", ""]));

let path = Uri::parse("foo:bar/baz")?.path();
assert_eq!(path, "bar/baz");
assert!(path.segments_if_absolute().is_none());

let path = Uri::parse("http://example.com")?.path();
assert!(path.is_empty());
assert!(path.segments_if_absolute().is_none());