Struct fluent_uri::Builder

pub struct Builder<S = UriStart> { /* private fields */ }

A builder for URI reference.

This struct is created by Uri::builder.

Examples

Basic usage:

use fluent_uri::{component::{Host, Scheme}, encoding::EStr, Uri};

let uri = Uri::builder()
    .scheme(Scheme::new("foo"))
    .authority(|b| {
        b.userinfo(EStr::new("user"))
            .host(Host::RegName(EStr::new("example.com")))
            .port(8042)
    })
    .path(EStr::new("/over/there"))
    .query(EStr::new("name=ferret"))
    .fragment(EStr::new("nose"))
    .build();

assert_eq!(
    uri.as_str(),
    "foo://user@example.com:8042/over/there?name=ferret#nose"
);

Only use EStr::new when you have a percent-encoded string at hand. You may otherwise encode and append data to an EString which derefs to EStr.

Constraints

Typestates are used to avoid misconfigurations, which puts the following constraints:

Components must be set from left to right, no repetition allowed.

Setting path is mandatory before calling build.

Methods userinfo, host, and port are only available within a call to authority.

Setting host is mandatory within a call to authority.

You may otherwise skip setting optional components (scheme, authority, userinfo, port, query, and fragment) with advance or set them optionally with optional.

The builder typestates are currently private. Please open an issue if it is a problem not being able to name the type of a builder.

Implementations

impl<S> Builder<S>

pub fn advance<T>(self) -> Builder<T>
where S: To<T>, T: AdvanceDst,

Advances the builder state, skipping optional components in between.

Variable rebinding may be necessary as this changes the type of the builder.

use fluent_uri::{component::{Host, Scheme}, encoding::EStr, Uri};

fn build(relative: bool) -> Uri<String> {
    let b = Uri::builder();
    let b = if relative {
        b.advance()
    } else {
        b.scheme(Scheme::new("http"))
            .authority(|b| b.host(Host::RegName(EStr::new("example.com"))))
    };
    b.path(EStr::new("/foo")).build()
}

assert_eq!(build(false).as_str(), "http://example.com/foo");
assert_eq!(build(true).as_str(), "/foo");

pub fn optional<F, V, T>(self, f: F, opt: Option<V>) -> Builder<T>
where F: FnOnce(Builder<S>, V) -> Builder<T>, S: To<T>, T: AdvanceDst,

Optionally calls a builder method with a value.

use fluent_uri::{encoding::EStr, Builder, Uri};

let uri = Uri::builder()
    .path(EStr::new("foo"))
    .optional(Builder::query, Some(EStr::new("bar")))
    .optional(Builder::fragment, None)
    .build();

assert_eq!(uri.as_str(), "foo?bar");

impl<S: To<SchemeEnd>> Builder<S>

pub fn scheme(self, scheme: &Scheme) -> Builder<SchemeEnd>

Sets the scheme component.

impl<S: To<AuthorityStart>> Builder<S>

pub fn authority<F, T>(self, f: F) -> Builder<AuthorityEnd>
where F: FnOnce(Builder<AuthorityStart>) -> Builder<T>, T: To<AuthorityEnd>,

Builds the authority component with the given function.

impl<S: To<UserinfoEnd>> Builder<S>

pub fn userinfo(self, userinfo: &EStr<Userinfo>) -> Builder<UserinfoEnd>

Sets the userinfo subcomponent of authority.

impl<S: To<HostEnd>> Builder<S>

pub fn host(self, host: Host<'_>) -> Builder<HostEnd>

Sets the host subcomponent of authority.

pub fn host_port_from_socket_addr<A: Into<SocketAddr>>( self, addr: A, default_port: u16 ) -> Builder<PortEnd>

Available on crate feature net only.

Sets the host and the port subcomponent of authority to the given socket address.

The port component is omitted when it equals the default value.

Examples

use fluent_uri::{encoding::EStr, Uri};
use std::net::{Ipv4Addr, Ipv6Addr, SocketAddrV4, SocketAddrV6};

let addr = SocketAddrV4::new(Ipv4Addr::LOCALHOST, 80);
let uri = Uri::builder()
    .authority(|b| b.host_port_from_socket_addr(addr, 80))
    .path(EStr::new(""))
    .build();
assert_eq!(uri.as_str(), "//127.0.0.1");

let uri = Uri::builder()
    .authority(|b| b.host_port_from_socket_addr(addr, 21))
    .path(EStr::new(""))
    .build();
assert_eq!(uri.as_str(), "//127.0.0.1:80");

let addr = SocketAddrV6::new(Ipv6Addr::LOCALHOST, 80, 0, 0);
let uri = Uri::builder()
    .authority(|b| b.host_port_from_socket_addr(addr, 80))
    .path(EStr::new(""))
    .build();
assert_eq!(uri.as_str(), "//[::1]");

impl<S: To<PortEnd>> Builder<S>

pub fn port<T: PortLike>(self, port: T) -> Builder<PortEnd>

Sets the port subcomponent of authority.

Takes either a u16 or &S where S: AsRef<str> + ?Sized as argument.

Panics

Panics if an input string is not a valid port according to Section 3.2.3 of RFC 3986.

pub fn port_with_default(self, port: u16, default: u16) -> Builder<PortEnd>

Sets the port subcomponent of authority, omitting it when it equals the default value.

impl<S: To<PathEnd>> Builder<S>

pub fn path(self, path: &EStr<Path>) -> Builder<PathEnd>

Sets the path component.

Panics

Panics if any of the following conditions is not met, as required in Section 3.3 of RFC 3986.

When authority is present, the path must either be empty or start with '/'.

When authority is absent, the path cannot start with "//".

In a relative-path reference, the first path segment cannot contain ':'.

impl<S: To<QueryEnd>> Builder<S>

pub fn query(self, query: &EStr<Query>) -> Builder<QueryEnd>

Sets the query component.

impl<S: To<FragmentEnd>> Builder<S>

pub fn fragment(self, fragment: &EStr<Fragment>) -> Builder<FragmentEnd>

Sets the fragment component.

impl<S: To<UriEnd>> Builder<S>

pub fn build(self) -> Uri<String>

Builds the URI reference.

Panics

Panics if the output length would be greater than u32::MAX.