Struct fluent_uri::Scheme

#[repr(transparent)]
pub struct Scheme(_);

The scheme component of URI reference.

Implementations

impl Scheme

pub fn as_str(&self) -> &str

Returns the scheme as a string slice.

Examples

use fluent_uri::Uri;

let uri = Uri::parse("HTTP://example.com/")?;
let scheme = uri.scheme().unwrap();
assert_eq!(scheme.as_str(), "HTTP");

pub fn to_lowercase(&self) -> String

Returns the scheme as a string in lower case.

Examples

use fluent_uri::Uri;

let uri = Uri::parse("HTTP://example.com/")?;
let scheme = uri.scheme().unwrap();
assert_eq!(scheme.to_lowercase(), "http");

pub fn eq_lowercase(&self, other: &str) -> bool

Checks if the scheme equals case-insensitively with a lowercase string.

This method is slightly faster than str::eq_ignore_ascii_case but will always return false if there is any uppercase letter in the given string.

Examples

use fluent_uri::Uri;

let uri = Uri::parse("HTTP://example.com/")?;
let scheme = uri.scheme().unwrap();
assert!(scheme.eq_lowercase("http"));
// Always return `false` if there's any uppercase letter in the given string.
assert!(!scheme.eq_lowercase("hTTp"));