Crate fluent_uri

A generic URI parser that strictly adheres to IETF RFC 3986.

See the documentation of Uri for more details.

Feature flags

All features except std are disabled by default. Note that the last two features each alter the enum HostData in a backward incompatible way that could make it impossible for two crates that depend on different features of fluent-uri to be used together.

std: Enables std support. This includes Error implementations and Ip{v4, v6}Addr support in HostData.

ipv_future: Enables the parsing of IPvFuture literal addresses, which fails with InvalidIpLiteral when disabled.

Only enable this feature when you have a compelling reason to do so, such as that you have to deal with an existing system where the IPvFuture format is in use.

rfc6874bis: Enables the parsing of IPv6 zone identifiers, such as in https://[fe80::abcd%en1].

This feature is based on the homonymous draft and is thus subject to change.

Modules

enc

Utilities for percent-encoding.

Structs

Authority

The authority component of URI reference.

Host

The host subcomponent of authority.

ParseError

An error occurred when parsing URI references.

Path

The path component of URI reference.

Scheme

The scheme component of URI reference.

Uri

A URI reference defined in RFC 3986.

View

A smart pointer that allows viewing a mutable byte slice as &T.

Enums

HostData

Structured host data.

ParseErrorKind

Detailed cause of a ParseError.