Struct fluent_uri::Uri

#[repr(C)]
pub struct Uri<T: Storage> { /* private fields */ }

A URI reference defined in RFC 3986.

Variants

There are three variants of Uri in total:

Uri<&str>: borrowed; immutable.

Uri<&mut [u8]>: borrowed; in-place mutable.

Uri<String>: owned; immutable.

Lifetimes are correctly handled in a way that Uri<&'a str> and Uri<&'a mut [u8]> both output references with lifetime 'a. This allows you to drop a temporary Uri while keeping the output references.

use fluent_uri::Uri;

let uri = Uri::parse("foo:bar").expect("invalid URI reference");
let path = uri.path();
drop(uri);
assert_eq!(path.as_str(), "bar");

Examples

Create and convert between Uri<&str> and Uri<String>.

use fluent_uri::Uri;

// Create a `Uri<&str>`.
let uri_a: Uri<&str> = Uri::parse("").expect("invalid URI reference");

// Create a `Uri<String>`.
let uri_b: Uri<String> = Uri::parse_from(String::new()).expect("invalid URI reference");

// Convert a `Uri<&str>` to a `Uri<String>`.
let uri_c: Uri<String> = uri_a.to_owned();

// Borrow a `Uri<String>` as a `Uri<&str>`.
let uri_d: &Uri<&str> = uri_b.borrow();

Decode path segments in-place and collect them into a Vec.

use fluent_uri::Uri;

fn decode_path_segments(uri: &mut [u8]) -> Option<Vec<&str>> {
    let mut uri = Uri::parse_mut(uri).ok()?;
    let segs = uri.take_path_mut().segments_mut();
    let mut out = Vec::new();
    for seg in segs {
        out.push(seg.decode_in_place().into_str().ok()?);
    }
    Some(out)
}
   
let mut uri = b"/path/to/my%20music".to_vec();
assert_eq!(decode_path_segments(&mut uri).unwrap(), ["path", "to", "my music"]);

Create a mutable copy of an immutable Uri in a buffer.

use fluent_uri::Uri;
use std::mem::MaybeUninit;

let uri = Uri::parse("https://www.rust-lang.org/").expect("invalid URI reference");
let mut buf = [MaybeUninit::uninit(); 256];
let uri_mut = uri
    .to_mut_in(&mut buf[..])
    .expect("buffer capacity overflow");

Implementations

impl<'a> Uri<&'a str>

pub fn parse<S: AsRef<[u8]> + ?Sized>(s: &S) -> Result<Uri<&str>, UriParseError>

Parses a URI reference from a byte sequence into a Uri<&str>.

This function validates the input strictly except that UTF-8 validation is not performed on a percent-encoded registered name (see Section 3.2.2, RFC 3986). Care should be taken when dealing with such cases.

Panics

Panics if the input length is greater than i32::MAX.

pub fn dup(&self) -> Uri<&'a str>

Duplicates this Uri<&str>.

pub fn to_owned(&self) -> Uri<String>

Creates a new Uri<String> by cloning the contents of this Uri<&str>.

impl<'i, 'o, T: Io<'i, 'o> + AsRef<str>> Uri<T>

pub fn as_str(&'i self) -> &'o str

Returns the URI reference as a string slice.

pub fn to_mut_in<'b, B: Buf + ?Sized>(
&self,
buf: &'b mut B
) -> Result<Uri<&'b mut [u8]>, B::ReserveError>

Creates a mutable copy of this Uri in the given buffer.

The type of a buffer may be:

Vec<u8>: grows automatically; bytes appended to the end.

[u8] or [MaybeUninit<u8>]: triggers a CapOverflowError when the capacity is too small; bytes written from the start.

impl<'i, 'o, T: Io<'i, 'o>> Uri<T>

pub fn scheme(&'i self) -> Option<&'o Scheme>

Returns the scheme component.

pub fn authority(&self) -> Option<&Authority<T>>

Returns the authority component.

pub fn path(&'i self) -> &'o Path

Returns the path component.

Panics

Panics if the path component is already taken.

pub fn query(&'i self) -> Option<&'o EStr>

Returns the query component.

pub fn fragment(&'i self) -> Option<&'o EStr>

Returns the fragment component.

pub fn is_relative(&self) -> bool

Returns true if the URI reference is relative, i.e., without a scheme.

Note that this function is not the opposite of is_absolute.

Examples

use fluent_uri::Uri;

let uri = Uri::parse("/path/to/file")?;
assert!(uri.is_relative());
let uri = Uri::parse("http://example.com/")?;
assert!(!uri.is_relative());

pub fn is_absolute(&self) -> bool

Returns true if the URI reference is absolute, i.e., with a scheme and without a fragment.

Note that this function is not the opposite of is_relative.

Examples

use fluent_uri::Uri;

let uri = Uri::parse("http://example.com/")?;
assert!(uri.is_absolute());
let uri = Uri::parse("http://example.com/#title1")?;
assert!(!uri.is_absolute());
let uri = Uri::parse("/path/to/file")?;
assert!(!uri.is_absolute());

impl<'a> Uri<&'a mut [u8]>

pub fn parse_mut<S: AsMut<[u8]> + ?Sized>(
s: &mut S
) -> Result<Uri<&mut [u8]>, UriParseError>

Parses a URI reference from a mutable byte sequence into a Uri<&mut [u8]>.

Panics

Panics if the input length is greater than i32::MAX.

pub fn into_mut_bytes(self) -> &'a mut [u8]

Consumes this Uri and yields the underlying mutable byte slice.

pub fn scheme_mut(&mut self) -> Option<&'a mut Scheme>

Returns the mutable scheme component.

pub fn take_authority_mut(&mut self) -> Option<AuthorityMut<'_, 'a>>

Takes the mutable authority component, leaving a None in its place.

pub fn take_path_mut(&mut self) -> PathMut<'a>

Takes the mutable path component.

Panics

Panics if the path component is already taken.

pub fn take_query_mut(&mut self) -> Option<EStrMut<'a>>

Takes the mutable query component, leaving a None in its place.

pub fn take_fragment_mut(&mut self) -> Option<EStrMut<'a>>

Takes the mutable fragment component, leaving a None in its place.

impl Uri<String>

pub fn parse_from<T: IntoOwnedUri>(
t: T
) -> Result<Uri<String>, (T, UriParseError)>

Parses a URI reference from a String or Vec<u8> into a Uri<String>.

Panics

Panics if the input capacity is greater than i32::MAX.

pub fn into_string(self) -> String

Consumes this Uri and yields the underlying String storage.

pub fn borrow(&self) -> &Uri<&str>

Borrows this Uri<String> as a reference to Uri<&str>.