Struct fluent_uri::Uri

pub struct Uri<T> { /* private fields */ }

A URI reference defined in RFC 3986.

Variants

Two variants of Uri are available: Uri<&str> (borrowed) and Uri<String> (owned).

Uri<&'a str> outputs references with lifetime 'a where possible:

use fluent_uri::Uri;

// Keep a reference to the path after dropping the `Uri`.
let path = Uri::parse("foo:bar")?.path();
assert_eq!(path, "bar");

Examples

Parse into and convert between Uri<&str> and Uri<String>:

use fluent_uri::Uri;

let s = "foo:bar";

// Parse into a `Uri<&str>` from a string slice.
let uri: Uri<&str> = Uri::parse(s)?;

// Parse into a `Uri<String>` from an owned string.
let uri_owned: Uri<String> = Uri::parse(s.to_owned()).map_err(|e| e.plain())?;

// When referencing a `Uri`, use `Uri<&str>`.
fn foo(uri: Uri<&str>) {
    // Convert a `Uri<&str>` to `Uri<String>`.
    let uri_owned: Uri<String> = uri.to_owned();
}

foo(uri);
// Borrow a `Uri<String>` as `Uri<&str>`.
foo(uri_owned.borrow());

See the documentation of Builder for examples of building a Uri from its components.

Implementations

impl<T> Uri<T>

pub fn parse<I>(input: I) -> Result<Self, I::Err>
where I: ToUri<Val = T>,

Parses a URI reference from a string into a Uri.

The return type is

Result<Uri<&str>, ParseError> for I = &str;

Result<Uri<String>, ParseError<String>> for I = String.

Returns Ok if and only if the string matches the URI-reference ABNF rule from RFC 3986.

You may recover an input String by calling ParseError::into_input.

Panics

Panics if the input length is greater than u32::MAX.

impl Uri<String>

pub fn builder() -> Builder

Creates a new builder for URI reference.

pub fn borrow(&self) -> Uri<&str>

Borrows this Uri<String> as Uri<&str>.

pub fn into_string(self) -> String

Consumes this Uri<String> and yields the underlying String.

impl Uri<&str>

pub fn to_owned(&self) -> Uri<String>

Creates a new Uri<String> by cloning the contents of this Uri<&str>.

impl<'i, 'o, T: BorrowOrShare<'i, 'o, str>> Uri<T>

pub fn as_str(&'i self) -> &'o str

Returns the URI reference as a string slice.

pub fn scheme(&'i self) -> Option<&'o Scheme>

Returns the scheme component.

pub fn authority(&self) -> Option<&Authority<T>>

Returns the authority component.

pub fn path(&'i self) -> &'o EStr<Path>

Returns the path component.

The returned EStr slice has extension methods for the path component.

pub fn query(&'i self) -> Option<&'o EStr<Query>>

Returns the query component.

pub fn fragment(&'i self) -> Option<&'o EStr<Fragment>>

Returns the fragment component.

pub fn is_relative_reference(&self) -> bool

Checks whether the URI reference is a relative reference, i.e., without a scheme.

Note that this method is not the opposite of is_absolute_uri.

Examples

use fluent_uri::Uri;

let uri = Uri::parse("/path/to/file")?;
assert!(uri.is_relative_reference());
let uri = Uri::parse("http://example.com/")?;
assert!(!uri.is_relative_reference());

pub fn is_absolute_uri(&self) -> bool

Checks whether the URI reference is an absolute URI, i.e., with a scheme and without a fragment.

Note that this method is not the opposite of is_relative_reference.

Examples

use fluent_uri::Uri;

let uri = Uri::parse("http://example.com/")?;
assert!(uri.is_absolute_uri());
let uri = Uri::parse("http://example.com/#title1")?;
assert!(!uri.is_absolute_uri());
let uri = Uri::parse("/path/to/file")?;
assert!(!uri.is_absolute_uri());