Struct fluent_uri::View

pub struct View<'a, T: ?Sized + Lens>(_, _);

A smart pointer that allows viewing a mutable byte slice as &T.

This struct was introduced considering the fact that a bare &mut EStr wouldn’t do for in-place decoding because such decoding breaks the invariant of EStr.

Six types of lenses may be used as T: EStr, str, Scheme, Authority, Host, and Path.

Implementations

impl<'a> View<'a, EStr>

An EStr view into a mutable byte slice that allows in-place percent-decoding.

pub fn decode_in_place(self) -> DecodeInPlace<'a>

Decodes the View<EStr> in-place.

pub fn split_view(self, delim: char) -> SplitView<'a>

Returns an iterator over subslices of the View<EStr> separated by the given delimiter.

Panics

Panics if the delimiter is not a reserved character.

pub fn split_once_view(self, delim: char) -> Result<(Self, Self), Self>

Splits the View<EStr> on the first occurrence of the given delimiter and returns prefix before delimiter and suffix after delimiter.

Returns Err(self) if the delimiter is not found.

Panics

Panics if the delimiter is not a reserved character.

impl<'a, T: ?Sized + Lens<Target = [u8]>> View<'a, T>

These methods are only available for lenses EStr, str, Scheme, and Path.

pub fn into_ref(self) -> &'a T

Consumes this View and yields the underlying &T.

pub fn into_bytes(self) -> &'a mut [u8]

Consumes this View and yields the underlying mutable byte slice.

impl<'a> View<'a, Scheme>

A Scheme view into a mutable byte slice that allows lowercasing in-place.

pub fn make_lowercase(&mut self)

Converts the scheme to lower case in-place.

Examples

use fluent_uri::Uri;

let mut bytes = *b"HTTP://example.com/";
let mut uri = Uri::parse_mut(&mut bytes)?;

let mut scheme = uri.take_scheme().unwrap();
scheme.make_lowercase();
assert_eq!(scheme.as_str(), "http");

impl<'i, 'a> View<'i, Authority<&'a mut [u8]>>

An Authority view into a mutable byte slice.

pub fn into_str_view(self) -> View<'a, str>

Consumes this View<Authority> and yields the underlying View<str>.

The userinfo or port subcomponent is truncated if it is already taken.

Panics

Panics if the host subcomponent is already taken.

pub fn take_userinfo(&mut self) -> Option<View<'a, EStr>>

Takes a view of the userinfo subcomponent, leaving a None in its place.

pub fn take_host(&mut self) -> View<'_, Host<&'a mut [u8]>>

Takes a view of the host subcomponent.

Panics

Panics if the host subcomponent is already taken.

pub fn take_port(&mut self) -> Option<View<'a, str>>

Takes a view of the port subcomponent, leaving a None in its place.

impl<'i, 'a> View<'i, Host<&'a mut [u8]>>

A Host view into a mutable byte slice.

pub fn into_str_view(self) -> View<'a, str>

Consumes this View<Host> and yields the underlying View<str>.

pub fn unwrap_reg_name(self) -> View<'a, EStr>

Consumes this View<Host> and yields the underlying View<EStr>, assuming that the host is a registered name.

Panics

Panics if the host is not a registered name.

impl<'a> View<'a, Path>

A Path view into a mutable byte slice.

pub fn into_estr_view(self) -> View<'a, EStr>

Consumes this View<Path> and yields the underlying View<EStr>.

pub fn segments_view(self) -> SplitView<'a>

Returns an iterator over the views of path segments.